Упростил код работы с индексами
