+++ StateInitialization/StateInitialization.Core/Actors/DisableIndexesActor.cs
@@ -11,17 +11,22 @@ namespace NuClear.StateInitialization.Core.Actors
 {
     public sealed class DisableIndexesActor<TDataObject> : IActor
     {
-        private readonly IActor _indexesManagementActor;
+        private readonly IndexManagementService _indexManagementService;
 
         public DisableIndexesActor(SqlConnection sqlConnection)
         {
-            _indexesManagementActor = new IndexesManagementActor<TDataObject>(sqlConnection);
+            _indexManagementService = new IndexManagementService(sqlConnection);
         }
 
         public IReadOnlyCollection<IEvent> ExecuteCommands(IReadOnlyCollection<ICommand> commands)
         {
             var disableCommand = commands.OfType<DisableIndexesCommand>().SingleOrDefault();
-            return disableCommand == null ? Array.Empty<IEvent>() : _indexesManagementActor.ExecuteCommands(new[] { disableCommand });
+            if (disableCommand != null)
+            {
+                _indexManagementService.DisableIndexes(disableCommand.MappingSchema, typeof(TDataObject));
+            }
+
+            return Array.Empty<IEvent>();
         }
     }
 }
