+++ StateInitialization/StateInitialization.Core/StateInitialization.Core.csproj
@@ -89,7 +89,7 @@
     <Compile Include="Actors\EnableIndexesActor.cs" />
     <Compile Include="Actors\DisableIndexesActor.cs" />
     <Compile Include="Actors\GlobalLocker.cs" />
-    <Compile Include="Actors\IndexesManagementActor.cs" />
+    <Compile Include="Actors\IndexManagementService.cs" />
     <Compile Include="Actors\TruncateTableActor.cs" />
     <Compile Include="Actors\SequentialPipelineActor.cs" />
     <Compile Include="Actors\ConstraintsManagementActor.cs" />
